OK-611 Haetaan lähtötiedot koostepalvelusta
