--- valintalaskenta-laskenta-service/src/main/java/fi/vm/sade/valinta/kooste/valintalaskenta/actor/LaskentaActor.java
@@ -1,7 +1,0 @@
-package fi.vm.sade.valinta.kooste.valintalaskenta.actor;
-
-public interface LaskentaActor {
-  void start();
-
-  void lopeta();
-}
