+++ valintalaskenta-laskenta-service/src/main/resources/application.properties
@@ -4,6 +4,7 @@ cas.service.suoritusrekisteri=https://${host.virkailija}/suoritusrekisteri/j_spr
 cas.service.ataru=https://${host.virkailija}/lomake-editori/auth/cas
 cas.service.valintapiste-service=https://${host.virkailija}/valintapiste-service/auth/cas
 cas.service.valintaperusteet-service=https://${host.virkailija}/valintaperusteet-service
+cas.service.koostepalvelu=https://${host.virkailija}/valintalaskentakoostepalvelu/j_spring_cas_security_check
 
 valintalaskentakoostepalvelu.valintaperusteet-service.baseurl=https://${host.virkailija}
