+++ valintalaskenta-laskenta-service/src/main/resources/valintalaskentakoostepalvelu-oph.properties
@@ -167,3 +167,5 @@ koski.massaluovutus.post=${baseurl-koski}/koski/api/massaluovutus
 # lisätty, TODO: varmista että tarvitaan
 valintalaskentakoostepalvelu.valintalaskenta-laskenta-service.baseurl=https://${host.virkailija}
 valintalaskentakoostepalvelu.valintaperusteet-service.baseurl=https://${host.virkailija}
+
+valintalaskentakoostepalvelu.lahtotiedot.baseurl=https://${host.virkailija}/valintalaskentakoostepalvelu/resources/valintalaskentakerralla/haku/$1/hakukohde/$2/lahtotiedot
